refactor folder structure
Change folder structure for more clarity
