+++ cea/optimization_new/containerclasses/supplySystemStructure.py
@@ -28,10 +28,10 @@ __status__ = "Production"
 
 
 import pandas as pd
-from cea.optimization_new.containerclasses.energysystems.energyCarrier import EnergyCarrier
-from cea.optimization_new.containerclasses.energysystems.energyFlow import EnergyFlow
+from cea.optimization_new.containerclasses.energyCarrier import EnergyCarrier
+from cea.optimization_new.containerclasses.energyFlow import EnergyFlow
 from cea.optimization_new.component import ActiveComponent, PassiveComponent
-from cea.optimization_new.containerclasses.optimization.capacityIndicator import CapacityIndicator, CapacityIndicatorVector
+from cea.optimization_new.helpercalsses.optimization.capacityIndicator import CapacityIndicator, CapacityIndicatorVector
 
 
 class SupplySystemStructure(object):
