+++ cea/optimization_new/helpercalsses/optimization/capacityIndicator.py
